+++ Lottery/src/Lottery/Data/RangeNumber.cs
@@ -2,7 +2,7 @@ namespace Lottery
 {
     public class RangeNumber : IRangeNumber
     {
-        public byte Value { get; set; }
+        public int Value { get; set; }
         public bool IsOdd { get; set; }
         public bool IsEven { get; set; }
         public bool IsUpper { get; set; }
